Implement Newton's method for finding roots of the function f(x) = x^3 - 3x^2 + 1. Added derivative approximation and error handling. User input for initial guess and error tolerance included.

+++ raizPorBiseccion.cpp
@@ -1,5 +1,7 @@
 #include <iostream>
+#include <iomanip> // Para setprecision
 #include <cmath>
+
 
 using namespace std;
 
@@ -10,7 +12,7 @@ double fu(double x)
 
 int main()
 {
-    float a = 0, b = 0, eMax = 0;
+    double a = 0, b = 0, eMax = 0;
     cout << "Ingrese el dominio minimo: ";
     cin >> a;
